Code clean-up

+++ src/ProcessingTools/Web/ProcessingTools.Web.Documents/Areas/Documents/Controllers/ArticlesController.cs
@@ -89,12 +89,9 @@ namespace ProcessingTools.Web.Documents.Areas.Documents.Controllers
         /// <param name="returnUrl">Return URL.</param>
         /// <returns><see cref="IActionResult"/>.</returns>
         [ActionName(IndexActionName)]
-        public async Task<IActionResult> Index(int? p, int? n, string returnUrl = null)
+        [System.Diagnostics.CodeAnalysis.SuppressMessage("Design", "CA1031:Do not catch general exception types", Justification = "Endpoint")]
+        public async Task<IActionResult> Index(int? p, int? n, Uri returnUrl = null)
         {
-            const string LogMessage = "Fetch Articles";
-
-            this.logger.LogTrace(LogMessage);
-
             int pageNumber = Math.Max(
                 PaginationConstants.MinimalPageNumber,
                 p ?? PaginationConstants.DefaultPageNumber);
@@ -114,7 +111,7 @@ namespace ProcessingTools.Web.Documents.Areas.Documents.Controllers
             catch (Exception ex)
             {
                 this.ModelState.AddModelError(string.Empty, ex.Message);
-                this.logger.LogError(ex, LogMessage);
+                this.logger.LogError(ex, string.Empty);
             }
 
             return this.View();
@@ -127,12 +124,9 @@ namespace ProcessingTools.Web.Documents.Areas.Documents.Controllers
         /// <returns><see cref="IActionResult"/>.</returns>
         [HttpGet]
         [ActionName(CreateActionName)]
-        public async Task<IActionResult> Create(string returnUrl = null)
+        [System.Diagnostics.CodeAnalysis.SuppressMessage("Design", "CA1031:Do not catch general exception types", Justification = "Endpoint")]
+        public async Task<IActionResult> Create(Uri returnUrl = null)
         {
-            const string LogMessage = "GET Create Article";
-
-            this.logger.LogTrace(LogMessage);
-
             try
             {
                 var viewModel = await this.service.GetArticleCreateViewModelAsync().ConfigureAwait(false);
@@ -143,7 +137,7 @@ namespace ProcessingTools.Web.Documents.Areas.Documents.Controllers
             catch (Exception ex)
             {
                 this.ModelState.AddModelError(string.Empty, ex.Message);
-                this.logger.LogError(ex, LogMessage);
+                this.logger.LogError(ex, string.Empty);
             }
 
             return this.View();
@@ -157,24 +151,21 @@ namespace ProcessingTools.Web.Documents.Areas.Documents.Controllers
         [HttpPost]
         [ValidateAntiForgeryToken]
         [ActionName(CreateActionName)]
+        [System.Diagnostics.CodeAnalysis.SuppressMessage("Design", "CA1031:Do not catch general exception types", Justification = "Endpoint")]
         public async Task<IActionResult> Create(ArticleCreateRequestModel model)
         {
-            const string LogMessage = "POST Create Article";
-
-            this.logger.LogTrace(LogMessage);
-
             try
             {
-                if (this.ModelState.IsValid)
+                if (model != null && this.ModelState.IsValid)
                 {
                     try
                     {
                         var ok = await this.service.CreateArticleAsync(model).ConfigureAwait(false);
                         if (ok)
                         {
-                            if (!string.IsNullOrWhiteSpace(model.ReturnUrl))
+                            if (model.ReturnUrl != null)
                             {
-                                return this.Redirect(model.ReturnUrl);
+                                return this.Redirect(model.ReturnUrl.ToString());
                             }
 
                             return this.RedirectToAction(IndexActionName);
@@ -185,19 +176,19 @@ namespace ProcessingTools.Web.Documents.Areas.Documents.Controllers
                     catch (Exception ex)
                     {
                         this.ModelState.AddModelError(string.Empty, ex.Message);
-                        this.logger.LogError(ex, LogMessage);
+                        this.logger.LogError(ex, string.Empty);
                     }
                 }
 
                 var viewModel = await this.service.MapToViewModelAsync(model).ConfigureAwait(false);
-                viewModel.ReturnUrl = model.ReturnUrl;
+                viewModel.ReturnUrl = model?.ReturnUrl;
 
                 return this.View(viewModel);
             }
             catch (Exception ex)
             {
                 this.ModelState.AddModelError(string.Empty, ex.Message);
-                this.logger.LogError(ex, LogMessage);
+                this.logger.LogError(ex, string.Empty);
             }
 
             return this.View();
@@ -210,12 +201,9 @@ namespace ProcessingTools.Web.Documents.Areas.Documents.Controllers
         /// <returns><see cref="IActionResult"/>.</returns>
         [HttpGet]
         [ActionName(CreateFromFileActionName)]
-        public async Task<IActionResult> CreateFromFile(string returnUrl = null)
+        [System.Diagnostics.CodeAnalysis.SuppressMessage("Design", "CA1031:Do not catch general exception types", Justification = "Endpoint")]
+        public async Task<IActionResult> CreateFromFile(Uri returnUrl = null)
         {
-            const string LogMessage = "GET CreateFromFile Article";
-
-            this.logger.LogTrace(LogMessage);
-
             try
             {
                 var viewModel = await this.service.GetArticleCreateFromFileViewModelAsync().ConfigureAwait(false);
@@ -226,7 +214,7 @@ namespace ProcessingTools.Web.Documents.Areas.Documents.Controllers
             catch (Exception ex)
             {
                 this.ModelState.AddModelError(string.Empty, ex.Message);
-                this.logger.LogError(ex, LogMessage);
+                this.logger.LogError(ex, string.Empty);
             }
 
             return this.View();
@@ -242,20 +230,17 @@ namespace ProcessingTools.Web.Documents.Areas.Documents.Controllers
         [HttpPost]
         [ValidateAntiForgeryToken]
         [ActionName(CreateFromFileActionName)]
-        public async Task<IActionResult> CreateFromFile(IFormFile file, string journalId, string returnUrl = null)
+        [System.Diagnostics.CodeAnalysis.SuppressMessage("Design", "CA1031:Do not catch general exception types", Justification = "Endpoint")]
+        public async Task<IActionResult> CreateFromFile(IFormFile file, string journalId, Uri returnUrl = null)
         {
-            const string LogMessage = "POST CreateFromFile Article";
-
-            this.logger.LogTrace(LogMessage);
-
             try
             {
                 var ok = await this.service.CreateFromFileArticleAsync(file, journalId).ConfigureAwait(false);
                 if (ok)
                 {
-                    if (!string.IsNullOrWhiteSpace(returnUrl))
+                    if (returnUrl != null)
                     {
-                        return this.Redirect(returnUrl);
+                        return this.Redirect(returnUrl.ToString());
                     }
 
                     return this.RedirectToAction(IndexActionName);
@@ -266,7 +251,7 @@ namespace ProcessingTools.Web.Documents.Areas.Documents.Controllers
             catch (Exception ex)
             {
                 this.ModelState.AddModelError(string.Empty, ex.Message);
-                this.logger.LogError(ex, LogMessage);
+                this.logger.LogError(ex, string.Empty);
             }
 
             var viewModel = await this.service.GetArticleCreateFromFileViewModelAsync().ConfigureAwait(false);
@@ -283,12 +268,9 @@ namespace ProcessingTools.Web.Documents.Areas.Documents.Controllers
         /// <returns><see cref="IActionResult"/>.</returns>
         [HttpGet]
         [ActionName(EditActionName)]
-        public async Task<IActionResult> Edit(string id, string returnUrl = null)
+        [System.Diagnostics.CodeAnalysis.SuppressMessage("Design", "CA1031:Do not catch general exception types", Justification = "Endpoint")]
+        public async Task<IActionResult> Edit(string id, Uri returnUrl = null)
         {
-            const string LogMessage = "GET Edit Article";
-
-            this.logger.LogTrace(LogMessage);
-
             try
             {
                 var viewModel = await this.service.GetArticleEditViewModelAsync(id).ConfigureAwait(false);
@@ -299,7 +281,7 @@ namespace ProcessingTools.Web.Documents.Areas.Documents.Controllers
             catch (Exception ex)
             {
                 this.ModelState.AddModelError(string.Empty, ex.Message);
-                this.logger.LogError(ex, LogMessage);
+                this.logger.LogError(ex, string.Empty);
             }
 
             return this.View();
@@ -313,24 +295,21 @@ namespace ProcessingTools.Web.Documents.Areas.Documents.Controllers
         [HttpPost]
         [ValidateAntiForgeryToken]
         [ActionName(EditActionName)]
+        [System.Diagnostics.CodeAnalysis.SuppressMessage("Design", "CA1031:Do not catch general exception types", Justification = "Endpoint")]
         public async Task<IActionResult> Edit(ArticleUpdateRequestModel model)
         {
-            const string LogMessage = "POST Edit Article";
-
-            this.logger.LogTrace(LogMessage);
-
             try
             {
-                if (this.ModelState.IsValid)
+                if (model != null && this.ModelState.IsValid)
                 {
                     try
                     {
                         var ok = await this.service.UpdateArticleAsync(model).ConfigureAwait(false);
                         if (ok)
                         {
-                            if (!string.IsNullOrWhiteSpace(model.ReturnUrl))
+                            if (model.ReturnUrl != null)
                             {
-                                return this.Redirect(model.ReturnUrl);
+                                return this.Redirect(model.ReturnUrl.ToString());
                             }
 
                             return this.RedirectToAction(EditActionName, new { model.Id });
@@ -341,19 +320,19 @@ namespace ProcessingTools.Web.Documents.Areas.Documents.Controllers
                     catch (Exception ex)
                     {
                         this.ModelState.AddModelError(string.Empty, ex.Message);
-                        this.logger.LogError(ex, LogMessage);
+                        this.logger.LogError(ex, string.Empty);
                     }
                 }
 
                 var viewModel = await this.service.MapToViewModelAsync(model).ConfigureAwait(false);
-                viewModel.ReturnUrl = model.ReturnUrl;
+                viewModel.ReturnUrl = model?.ReturnUrl;
 
                 return this.View(viewModel);
             }
             catch (Exception ex)
             {
                 this.ModelState.AddModelError(string.Empty, ex.Message);
-                this.logger.LogError(ex, LogMessage);
+                this.logger.LogError(ex, string.Empty);
             }
 
             return this.View();
@@ -367,12 +346,9 @@ namespace ProcessingTools.Web.Documents.Areas.Documents.Controllers
         /// <returns><see cref="IActionResult"/>.</returns>
         [HttpGet]
         [ActionName(DeleteActionName)]
-        public async Task<IActionResult> Delete(string id, string returnUrl = null)
+        [System.Diagnostics.CodeAnalysis.SuppressMessage("Design", "CA1031:Do not catch general exception types", Justification = "Endpoint")]
+        public async Task<IActionResult> Delete(string id, Uri returnUrl = null)
         {
-            const string LogMessage = "GET Delete Article";
-
-            this.logger.LogTrace(LogMessage);
-
             try
             {
                 var viewModel = await this.service.GetArticleDeleteViewModelAsync(id).ConfigureAwait(false);
@@ -383,7 +359,7 @@ namespace ProcessingTools.Web.Documents.Areas.Documents.Controllers
             catch (Exception ex)
             {
                 this.ModelState.AddModelError(string.Empty, ex.Message);
-                this.logger.LogError(ex, LogMessage);
+                this.logger.LogError(ex, string.Empty);
             }
 
             return this.View();
@@ -397,24 +373,21 @@ namespace ProcessingTools.Web.Documents.Areas.Documents.Controllers
         [HttpPost]
         [ValidateAntiForgeryToken]
         [ActionName(DeleteActionName)]
+        [System.Diagnostics.CodeAnalysis.SuppressMessage("Design", "CA1031:Do not catch general exception types", Justification = "Endpoint")]
         public async Task<IActionResult> Delete(ArticleDeleteRequestModel model)
         {
-            const string LogMessage = "POST Delete Article";
-
-            this.logger.LogTrace(LogMessage);
-
             try
             {
-                if (this.ModelState.IsValid)
+                if (model != null && this.ModelState.IsValid)
                 {
                     try
                     {
                         var ok = await this.service.DeleteArticleAsync(model.Id).ConfigureAwait(false);
                         if (ok)
                         {
-                            if (!string.IsNullOrWhiteSpace(model.ReturnUrl))
+                            if (model.ReturnUrl != null)
                             {
-                                return this.Redirect(model.ReturnUrl);
+                                return this.Redirect(model.ReturnUrl.ToString());
                             }
 
                             return this.RedirectToAction(IndexActionName);
@@ -425,19 +398,19 @@ namespace ProcessingTools.Web.Documents.Areas.Documents.Controllers
                     catch (Exception ex)
                     {
                         this.ModelState.AddModelError(string.Empty, ex.Message);
-                        this.logger.LogError(ex, LogMessage);
+                        this.logger.LogError(ex, string.Empty);
                     }
                 }
 
                 var viewModel = await this.service.MapToViewModelAsync(model).ConfigureAwait(false);
-                viewModel.ReturnUrl = model.ReturnUrl;
+                viewModel.ReturnUrl = model?.ReturnUrl;
 
                 return this.View(viewModel);
             }
             catch (Exception ex)
             {
                 this.ModelState.AddModelError(string.Empty, ex.Message);
-                this.logger.LogError(ex, LogMessage);
+                this.logger.LogError(ex, string.Empty);
             }
 
             return this.View();
@@ -450,20 +423,17 @@ namespace ProcessingTools.Web.Documents.Areas.Documents.Controllers
         /// <param name="returnUrl">Return URL.</param>
         /// <returns><see cref="IActionResult"/>.</returns>
         [ActionName(FinalizeActionName)]
-        public async Task<IActionResult> Finalize(string id, string returnUrl = null)
+        [System.Diagnostics.CodeAnalysis.SuppressMessage("Design", "CA1031:Do not catch general exception types", Justification = "Endpoint")]
+        public async Task<IActionResult> Finalize(string id, Uri returnUrl = null)
         {
-            const string LogMessage = "Finalize Article";
-
-            this.logger.LogTrace(LogMessage);
-
             try
             {
                 var ok = await this.service.FinalizeArticleAsync(id).ConfigureAwait(false);
                 if (ok)
                 {
-                    if (!string.IsNullOrWhiteSpace(returnUrl))
+                    if (returnUrl != null)
                     {
-                        return this.Redirect(returnUrl);
+                        return this.Redirect(returnUrl.ToString());
                     }
 
                     return this.RedirectToAction(IndexActionName);
@@ -471,7 +441,7 @@ namespace ProcessingTools.Web.Documents.Areas.Documents.Controllers
             }
             catch (Exception ex)
             {
-                this.logger.LogError(ex, LogMessage);
+                this.logger.LogError(ex, string.Empty);
             }
 
             return this.RedirectToAction(IndexActionName);
@@ -484,12 +454,9 @@ namespace ProcessingTools.Web.Documents.Areas.Documents.Controllers
         /// <param name="returnUrl">Return URL.</param>
         /// <returns><see cref="IActionResult"/>.</returns>
         [ActionName(DetailsActionName)]
-        public async Task<IActionResult> Details(string id, string returnUrl = null)
+        [System.Diagnostics.CodeAnalysis.SuppressMessage("Design", "CA1031:Do not catch general exception types", Justification = "Endpoint")]
+        public async Task<IActionResult> Details(string id, Uri returnUrl = null)
         {
-            const string LogMessage = "Fetch Article Details";
-
-            this.logger.LogTrace(LogMessage);
-
             try
             {
                 var viewModel = await this.service.GetArticleDetailsViewModelAsync(id).ConfigureAwait(false);
@@ -500,7 +467,7 @@ namespace ProcessingTools.Web.Documents.Areas.Documents.Controllers
             catch (Exception ex)
             {
                 this.ModelState.AddModelError(string.Empty, ex.Message);
-                this.logger.LogError(ex, LogMessage);
+                this.logger.LogError(ex, string.Empty);
             }
 
             return this.View();
@@ -513,12 +480,9 @@ namespace ProcessingTools.Web.Documents.Areas.Documents.Controllers
         /// <param name="returnUrl">Return URL.</param>
         /// <returns><see cref="IActionResult"/>.</returns>
         [ActionName(DocumentsActionName)]
-        public async Task<IActionResult> Documents(string id, string returnUrl = null)
+        [System.Diagnostics.CodeAnalysis.SuppressMessage("Design", "CA1031:Do not catch general exception types", Justification = "Endpoint")]
+        public async Task<IActionResult> Documents(string id, Uri returnUrl = null)
         {
-            const string LogMessage = "Fetch Article Documents";
-
-            this.logger.LogTrace(LogMessage);
-
             try
             {
                 var viewModel = await this.service.GetArticleDocumentsViewModelAsync(id).ConfigureAwait(false);
@@ -529,7 +493,7 @@ namespace ProcessingTools.Web.Documents.Areas.Documents.Controllers
             catch (Exception ex)
             {
                 this.ModelState.AddModelError(string.Empty, ex.Message);
-                this.logger.LogError(ex, LogMessage);
+                this.logger.LogError(ex, string.Empty);
             }
 
             return this.View();
